requirements.txt
Flask
pymysql
SQLAlchemy
pandas
selenium
webdriver-manager
cryptography
mysql-connector-python
